add top_p

+++ skythought/skythought_evals/util/metrics.py
@@ -13,7 +13,7 @@ def _pass_at_k(n, c, k):
     """
     if n - c < k:
         return 1.0
-    return 1.0 - np.prod(1.0 - k / np.arange(n - c + 1, n + 1))
+    return float(1.0 - np.prod(1.0 - k / np.arange(n - c + 1, n + 1)))
 
 
 def pass_at_k(N: int, temp_to_scores: Dict[str, Dict[str, Any]]):
